Moved kernel to force full 64-bit support for things like integer wme values, timetags, stats and counters, rete data structures.
General notes:
- Motivation for this is support for long-running agents. We're already rolling over stats numbers and timetags.
- Arguably this should be on a branch, but it is easier to check in and force testing by just putting it on the trunk. I'll roll it back and move it to a branch if there are problems or sufficient resistance.
- This check-in only compiles on and is tested on Windows. Fixes for linux and mac will be in the next couple check-ins.
- The "long long" workaround for SWIG (see below) may not work right on linux/mac and, if it doesn't, I'll roll back this change off the trunk and put it on a branch until we figure out a better workaround for SWIG and int64_t.

Details:
- Removal of hackish usage of intptr_t and uintptr_t as a flexible-size integer field. Removed associated data structures in soar_db.
- "long" and "unsigned long" are 64-bits on GCC on linux, mac; only 32-bits on MSVC. Changed these fields to the C99 stdint.h standard int64_t and uint64_t throughout the kernel and SML so that all platforms now see them as 64-bits. This could potentially seriously impact performance on 32-bit platforms.
- SWIG (even the newest version) currently interprets int64_t as some kind of pointer type on the hardware I've tested it on so far. In order to fix this, I changed instances of int64_t to "long long" which it correctly converts. This might not work on GCC.
- There are still plenty of 32-bit integers in the kernel and SML, the main changes were to stats, int wme values, timetags, rete data structures, tc numbers.
- The ClientSML interface was added to so that the change is backwards-compatible when compiled with reasonable warning levels. New functions have been added for those that need to return 64-bit values and are incapable of overloading (such as IntElement::GetValue() which takes no arguments). These functions are named the same with a 64 appended to the name (IntElement::GetValue64()). 
- Rete net loading and saving updated so that the new 64-bit rete data structures can be saved with full precision. This is implemented as a new rete net file version (4 instead of 3). The old 32-bit rete-nets (version 3) can still be loaded with no changes. The rete-net command should probably take a new argument to enable 32-bit saves (loads are auto-detected). I've tested this locally but have not made a cppunit test for it yet.
- The random number generator is 32-bit, but was changed to use stdint.h instead of it's own 32-bit type.
- There is no 64-bit version of the SML tag for the integer type. These are used when passing results and arguments and such back and forth through SML. These integers can be referenced as 32-bits or 64-bits as necessary, they are stored in signed 64-bit fields on the structures or as strings of arbitrary precision when rendered to XML.

+++ CLI/src/cli_chunknameformat.cpp
@@ -35,7 +35,7 @@ bool CommandLineInterface::ParseChunkNameFormat(std::vector<std::string>& argv) 
 
 	bool changeFormat = false;
 	bool countFlag = false;
-	int count = -1;
+	int64_t count = -1;
 	bool patternFlag = false;
 	std::string pattern;
 	bool longFormat = true;
@@ -76,7 +76,7 @@ bool CommandLineInterface::ParseChunkNameFormat(std::vector<std::string>& argv) 
 	return DoChunkNameFormat(changeFormat ? &longFormat : 0, countFlag ? &count : 0, patternFlag ? &pattern : 0);
 }
 
-bool CommandLineInterface::DoChunkNameFormat(const bool* pLongFormat, const int* pCount, const std::string* pPrefix) {
+bool CommandLineInterface::DoChunkNameFormat(const bool* pLongFormat, const int64_t* pCount, const std::string* pPrefix) {
 	if (!pLongFormat && !pCount && !pPrefix) {
 		if (m_RawOutput) {
 			m_Result << "Using " << (m_pAgentSoar->sysparams[USE_LONG_CHUNK_NAMES] ? "long" : "short") << " chunk format.";
@@ -91,8 +91,8 @@ bool CommandLineInterface::DoChunkNameFormat(const bool* pLongFormat, const int*
 	if (pCount) {
 		if (*pCount >= 0) {
 			if (*pCount >= m_pAgentSoar->sysparams[MAX_CHUNKS_SYSPARAM]) return SetError(CLIError::kCountGreaterThanMaxChunks);
-			if (static_cast<unsigned long>(*pCount) < m_pAgentSoar->chunk_count ) return SetError(CLIError::kCountLessThanChunks);
-			m_pAgentSoar->chunk_count = *pCount;
+			if (static_cast<uint64_t>(*pCount) < m_pAgentSoar->chunk_count ) return SetError(CLIError::kCountLessThanChunks);
+			m_pAgentSoar->chunk_count = static_cast<uint64_t>(*pCount);
 		} else {
 			// query
 			if (m_RawOutput) {
