+++ ConnectionSML/src/thread_OSspecific.h
@@ -36,7 +36,7 @@ class OSSpecificEvent
 public:
 	virtual ~OSSpecificEvent() {} ;
 	virtual void WaitForEventForever() = 0 ;
-	virtual bool WaitForEvent(long seconds, long milliseconds) = 0 ;
+	virtual bool WaitForEvent(int seconds, int milliseconds) = 0 ;
 	virtual void TriggerEvent() = 0 ;
 } ;
